backend/requirements.txt
fastapi
uvicorn
pydantic[email]
bcrypt
slowapi
python-dotenv
pymongo
email-validator
